requirements.txt
tqdm == 4.65.0
flask == 2.3.2
Pillow == 10.0.0
numpy == 1.23.5
pandas == 2.0.3
torchvision == 0.15.2
torch == 2.0.1
timm == 0.9.12
scikit-learn == 1.3.0